Refactor speakers component and desktop filter
