+++ src/pages/speakers.tsx
@@ -1,4 +1,4 @@
-import React, { ReactElement } from 'react';
+import React, { ReactElement, useState } from 'react';
 
 import { useStaticQuery, graphql } from 'gatsby';
 
@@ -7,69 +7,60 @@ import SEO from '../components/shared/seo';
 
 import Speakers from '../speakers/components/speakers';
 
-import { SpeakersFilterProvider } from '../speakers/contexts/filters';
+import { SpeakersFilter } from '../speakers/contexts/filters';
+import { FluidImage } from '../constants/prop-types';
+import { ALL_STREAMS } from '../constants/app';
+import { Filter } from '../entities/entities';
+
+interface Action {
+    type: string;
+    payload: Filter;
+}
+
+const INIT_STATE: Filter = {
+    eventType: ALL_STREAMS,
+    searchStr: '',
+};
 
 const SpeakersPage = (): ReactElement => {
+
     const data = useStaticQuery(graphql`
-        {
-            markdownRemark(frontmatter: { templateKey: { eq: "speakers-page" } }) {
-                frontmatter {
-                    pageTitle
-                }
+    {
+        markdownRemark(frontmatter: { templateKey: { eq: "speakers-page" } }) {
+            frontmatter {
+                pageTitle
             }
-            allMarkdownRemark(
-                filter: { fields: { collection: { eq: "speakers" } } }
-                sort: { fields: [frontmatter___name], order: ASC }
-            ) {
-                edges {
-                    node {
-                        id
-                        frontmatter {
-                            name
-                            company
-                            jobTitle
-                            photo {
-                                childImageSharp {
-                                    fluid(maxWidth: 400) {
-                                        ...GatsbyImageSharpFluid
-                                    }
-                                }
-                            }
-                            streams
-                            socialNetworks {
-                                type
-                                url
-                            }
-                        }
-                    }
-                }
-            }
-        }
+        }    
+    }
     `);
 
-    const speakers = data.allMarkdownRemark.edges.map(({ node }) => {
-        const {
-            frontmatter: { name, company, jobTitle, socialNetworks, photo, streams },
-            id,
-        } = node;
+    const speakersFilterReducer = (state: Filter, action: Action): Filter => {
+        switch (action.type) {
+            case 'setEventType': {
+                return { ...state, eventType: action.payload.eventType };
+            }
+            case 'setSearchStr': {
+                return { ...state, searchStr: action.payload.searchStr };
+            }
+            case 'setFilters': {
+                return { ...state, searchStr: action.payload.searchStr, eventType: action.payload.eventType };
+            }
+            case 'reset': {
+                return { ...INIT_STATE };
+            }
+            default: {
+                throw new Error(`Unhandled action type: ${action.type}`);
+            }
+        }
+    };
 
-        return {
-            id,
-            name,
-            company,
-            jobTitle,
-            socialNetworks,
-            photo,
-            streams,
-        };
-    });
+    const [filter, setFilter] = useState(INIT_STATE);
 
     return (
         <Layout>
             <SEO title={data.markdownRemark.frontmatter.pageTitle} />
-            <SpeakersFilterProvider>
-                <Speakers speakers={speakers} />
-            </SpeakersFilterProvider>
+            <SpeakersFilter setFilter={setFilter} />
+            <Speakers filter={filter}/>
         </Layout>
     );
 };
