+++ src/speakers/components/desktop-filters.tsx
@@ -1,113 +1,44 @@
-import React, { useCallback } from 'react';
-import classNames from 'classnames';
+import React, { useCallback, ReactElement, Dispatch, useState, useEffect } from 'react';
 
 import Box from '@material-ui/core/Box';
-import { makeStyles } from '@material-ui/core/styles';
 import MenuItem from '@material-ui/core/MenuItem';
 import Select from '@material-ui/core/Select';
 import InputBase from '@material-ui/core/InputBase';
 import InputLabel from '@material-ui/core/InputLabel';
 
 import { Streams, ALL_STREAMS } from '../../constants/app';
+import { Filter } from '../../entities/entities';
+import './desktop-filters.scss';
 
-import { useSpeakersFilterState, useSpeakersFiltersDispatch } from '../contexts/filters';
-
-const useStyles = makeStyles(() => ({
-    filterContainer: {
-        display: 'flex',
-        boxShadow: '5px 5px 20px #D1D1D1',
-        borderRadius: '10px',
-        margin: '20px 0',
-        position: 'sticky',
-        top: '70px',
-        width: '100%',
-        zIndex: '10',
-        background: '#fff',
-        padding: '15px 20px',
-    },
-    searchOptionText: {
-        fontSize: '14px',
-        lineHeight: '16px',
-    },
-    searchItemLabel: {
-        textTransform: 'uppercase',
-        color: '#000000',
-        marginRight: '10px',
-        flexShrink: '0',
-    },
-    searchByNameContainer: {
-        display: 'flex',
-        alignItems: 'center',
-        flexGrow: '1',
-        marginRight: '40px',
-    },
-    searchInputWrapper: {
-        background: '#FFFFFF',
-        border: '1px solid #EFEFEF',
-        borderRadius: '5px',
-    },
-    streamSelectInput: {
-        padding: '12px 10px',
-        minWidth: '150px',
-        textTransform: 'uppercase',
-    },
-    searchByNameInput: {
-        padding: '12px 10px',
-    },
-    dropdownItem: {
-        textTransform: 'uppercase',
-        fontSize: '14px',
-        lineHeight: '16px',
-    },
-    filterText: {
-        fontSize: '14px',
-        lineHeight: '16px',
-        color: '#6D7278',
-    },
-    '@media (max-width: 800px)': {
-        filterContainer: {
-            flexDirection: 'column',
-        },
-        searchByNameContainer: {
-            margin: '0 0 10px 0',
-        },
-    },
-}));
-
-const DesktopFilters = (): ReactElement => {
-    const classes = useStyles();
-
-    const { eventType, searchStr } = useSpeakersFilterState();
-
-    const filtersDispatch = useSpeakersFiltersDispatch();
+const DesktopFilters = ({ setFilter }: { setFilter: Dispatch<Filter> }): ReactElement => {
+    const [eventType, setEventType] = useState(ALL_STREAMS);
+    const [searchStr, setSearchStr] = useState('');
 
     const handleSearchStrChange = useCallback(event => {
-        filtersDispatch({
-            type: 'setSearchStr',
-            payload: { searchStr: event.target.value },
-        });
+        setSearchStr(event.target.value as string);
     }, []);
 
     const handleEventTypeChange = useCallback(event => {
-        filtersDispatch({
-            type: 'setEventType',
-            payload: { eventType: event.target.value },
-        });
+        setEventType(event.target.value as string);
     }, []);
 
+    useEffect(() => {
+        setFilter({ searchStr: searchStr, eventType: eventType });
+    }, [searchStr, eventType]);
+
     return (
-        <Box className={classes.filterContainer}>
-            <Box className={classes.searchByNameContainer}>
+        <Box className='filterContainer'>
+            <Box className='searchByNameContainer'>
                 <InputLabel
-                    className={classNames(classes.searchItemLabel, classes.searchOptionText)}
+                    className='searchItemLabel searchOptionText'
                     htmlFor="searchByNameInput"
                 >
                     speaker full name
                 </InputLabel>
                 <InputBase
                     id="searchByNameInput"
-                    className={classNames(classes.searchInputWrapper, classes.searchOptionText)}
-                    classes={{ input: classes.searchByNameInput }}
+                    className='searchInputWrapper searchOptionText'
+                    classes={{ input: 'searchByNameInput' }}
                     placeholder="Type any words to start search"
                     onChange={handleSearchStrChange}
                     value={searchStr}
@@ -118,7 +49,7 @@ const DesktopFilters = (): ReactElement => {
 
             <Box display="flex" alignItems="center">
                 <InputLabel
-                    className={classNames(classes.searchItemLabel, classes.searchOptionText)}
+                    className='searchItemLabel searchOptionText'
                     htmlFor="searchByStreamSelect"
                 >
                     Stream
@@ -130,22 +61,22 @@ const DesktopFilters = (): ReactElement => {
                     onChange={handleEventTypeChange}
                     input={
                         <InputBase
-                            className={classNames(classes.searchInputWrapper, classes.searchOptionText)}
-                            classes={{ input: classes.streamSelectInput }}
+                            className='searchInputWrapper searchOptionText'
+                            classes={{ input: 'streamSelectInput' }}
                         />
                     }
                     fullWidth
                 >
-                    <MenuItem classes={{ root: classes.dropdownItem }} value={ALL_STREAMS}>
+                    <MenuItem classes={{ root: 'dropdownItem' }} value={ALL_STREAMS}>
                         All
                     </MenuItem>
-                    <MenuItem classes={{ root: classes.dropdownItem }} value={Streams.WEB}>
+                    <MenuItem classes={{ root: 'dropdownItem' }} value={Streams.WEB}>
                         Web Meetup
                     </MenuItem>
-                    <MenuItem classes={{ root: classes.dropdownItem }} value={Streams.MOBILE}>
+                    <MenuItem classes={{ root: 'dropdownItem' }} value={Streams.MOBILE}>
                         Mobile Meetup
                     </MenuItem>
-                    <MenuItem classes={{ root: classes.dropdownItem }} value={Streams.CLOUD}>
+                    <MenuItem classes={{ root: 'dropdownItem' }} value={Streams.CLOUD}>
                         Cloud Meetup
                     </MenuItem>
                 </Select>
