+++ src/speakers/contexts/filters.tsx
@@ -1,76 +1,29 @@
-import React, { useReducer, createContext, useContext, ReactElement, Dispatch } from 'react';
+import React, { useReducer, ReactElement, Dispatch, SetStateAction } from 'react';
 
 import { ALL_STREAMS } from '../../constants/app';
+import { Box, Hidden } from '@material-ui/core';
+import DesktopFilters from '../components/desktop-filters';
+import MobileFilters from '../components/mobile-filters';
 
 interface Filter {
     eventType: string;
     searchStr: string;
 }
 
-interface Action {
-    type: string;
-    payload: Filter;
-}
-
-const INIT_STATE: Filter = {
-    eventType: ALL_STREAMS,
-    searchStr: '',
-};
-
-const SpeakersFilterStateContext = createContext<Filter>(INIT_STATE);
-const SpeakersFilterDispatchContext = createContext<Dispatch<Action>>(undefined);
-
-const speakersFilterReducer = (state: Filter, action: Action): Filter => {
-    switch (action.type) {
-        case 'setEventType': {
-            return { ...state, eventType: action.payload.eventType };
-        }
-        case 'setSearchStr': {
-            return { ...state, searchStr: action.payload.searchStr };
-        }
-        case 'setFilters': {
-            return { ...state, searchStr: action.payload.searchStr, eventType: action.payload.eventType };
-        }
-        case 'reset': {
-            return { ...INIT_STATE };
-        }
-        default: {
-            throw new Error(`Unhandled action type: ${action.type}`);
-        }
-    }
-};
-
-interface Props {
-    children?: ReactElement;
-}
-
-const SpeakersFilterProvider = ({ children }: Props): ReactElement => {
-    const [state, dispatch] = useReducer(speakersFilterReducer, INIT_STATE);
+const SpeakersFilter = ({setFilter }: {setFilter: Dispatch<Filter> }): ReactElement => {
+    //const [state, dispatch] = useReducer(speakersFilterReducer, INIT_STATE);
 
     return (
-        <SpeakersFilterStateContext.Provider value={state}>
-            <SpeakersFilterDispatchContext.Provider value={dispatch}>{children}</SpeakersFilterDispatchContext.Provider>
-        </SpeakersFilterStateContext.Provider>
+        <Box className="filterWrapper">
+            <Hidden xsDown>
+                <DesktopFilters setFilter={setFilter} />
+            </Hidden>
+
+            <Hidden smUp>
+                <MobileFilters />
+            </Hidden>
+        </Box>
     );
 };
 
-const useSpeakersFilterState = (): Filter => {
-    const context = useContext(SpeakersFilterStateContext);
-
-    if (context === undefined) {
-        throw new Error('useSpeakersFilterState must be used within a SpeakersFilterProvider');
-    }
-
-    return context;
-};
-
-const useSpeakersFiltersDispatch = (): Dispatch<Action> => {
-    const context = useContext(SpeakersFilterDispatchContext);
-
-    if (context === undefined) {
-        throw new Error('useSpeakersFiltersDispatch must be used within a SpeakersFilterProvider');
-    }
-    return context;
-};
-
-export { SpeakersFilterProvider, useSpeakersFilterState, useSpeakersFiltersDispatch };
+export { SpeakersFilter };
